Added utilities and materials files, split out related functions
